src/app/login-main/login-main.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';

import { LoginMainComponent } from './login-main.component';

describe('LoginMainComponent', () => {
  let component: LoginMainComponent;
  let fixture: ComponentFixture<LoginMainComponent>;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      declarations: [ LoginMainComponent ]
    })
    .compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(LoginMainComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should create the app', () => {
    const fixture = TestBed.createComponent(LoginMainComponent);
    const app = fixture.componentInstance;
    expect(app).toBeTruthy();
  });

  it(`should have as title 'Expense Reimbursement System'`, () => {
    const fixture = TestBed.createComponent(LoginMainComponent);
    const app = fixture.componentInstance;
    expect(app.title).toEqual('Expense Reimbursement System');
  });

  it('should render title', () => {
    const fixture = TestBed.createComponent(LoginMainComponent);
    fixture.detectChanges();
    const compiled = fixture.nativeElement as HTMLElement;
    expect(compiled.querySelector('.container-fluid h1')?.textContent).toContain('Expense Reimbursement System');
  });

});
